+++ source/extensions/omni.isaac.quadruped/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.4.3] - 2024-02-12
+### Changed
+- Removed IMU sensor from unitree.py (since the controller uses ground truth data)
+- Reduced the frequency of osqp solver from every physics step to every 5 physics steps
+- Contact sensor now uses the interface instead directly of the python wrapper
+
 ## [1.4.2] - 2023-02-02
 ### Changed
 - Updated path to the nucleus extension